Move `getBlockNameFromUniqueID`
