+++ src/extensions/attributes/index.js
@@ -1,12 +1,13 @@
 import './withAttributes';
 
+export { default as generateStyleID } from './generateStyleID';
 export { default as getBlockData } from './getBlockData';
+export { default as getBlockNameFromUniqueID } from './getBlockNameFromUniqueID';
 export { default as getClientIdFromUniqueId } from './getClientIdFromUniqueId';
 export { default as getRowGapProps } from './getRowGapProps';
 export { default as getUpdatedBGLayersWithNewUniqueID } from './getUpdatedBGLayersWithNewUniqueID';
 export { default as getUpdatedSVGDataAndElement } from './getUpdatedSVGDataAndElement';
 export { default as handleOnReset } from './handleOnReset';
-export { default as generateStyleID } from './generateStyleID';
+export { default as uniqueCustomLabelGenerator } from './uniqueCustomLabelGenerator';
 export { default as uniqueIDGenerator } from './uniqueIDGenerator';
 export { default as uniqueIDRemover } from './uniqueIDRemover';
-export { default as uniqueCustomLabelGenerator } from './uniqueCustomLabelGenerator';
