+++ src/extensions/styles/migrators/utils.js
@@ -1,19 +1,7 @@
 /**
  * Internal dependencies
  */
-import { getBlockData } from '../../attributes';
-
-/**
- * Returns block name from uniqueID
- *
- * @example uniqueID: accordion-maxi-123 -> accordion-maxi
- * @example uniqueID: accordion-maxi-1se8ef1z-u -> accordion-maxi
- */
-export const getBlockNameFromUniqueID = uniqueID => {
-	const match = uniqueID.match(/^(.*?)(-\d+|-[\w\d]+-u)$/);
-	if (match) return match[1];
-	return uniqueID; // fallback
-};
+import { getBlockData, getBlockNameFromUniqueID } from '../../attributes';
 
 /**
  * Returns block data from uniqueID
